refactor(marketplace): 拆分密钥迁移与连接器子模块
