+++ pinvou3-app/src-tauri/src/features/marketplace/mod.rs
@@ -3,6 +3,21 @@
 //! 每个工具是一个 MCP server，元数据定义在 `manifest.json`。
 //! 安装状态持久化在 `~/.pinvou3/marketplace/installed.json`。
 //! 安装/卸载时同步修改 `~/.pinvou3/bundle/mcp.json`。
+//!
+//! 本模块是 facade:把原本 2600+ 行的 god-module 按职责拆成子模块,
+//! 对外 pub 面通过 `pub use` 保持不变。
+//!
+//! - [`types`]      — manifest/info/迁移结果等数据类型
+//! - [`secrets`]    — 密钥/凭证助手 + MarketplaceManager 的 secret 读写方法
+//! - [`validation`] — 远程 MCP 连接校验
+//! - [`migration`]  — mcp.json 旧版明文密钥迁移
+//! - [`connectors`] — connector 注册/注销(含拆分后的 add_to_mcp_json remote/local 分支)
+
+mod connectors;
+mod migration;
+mod secrets;
+mod types;
+mod validation;
 
 pub mod bundle;
 pub mod skill_marketplace;
@@ -10,15 +25,16 @@ pub mod skill_scope;
 
 use std::path::PathBuf;
 
-use deepseek_tui::mcp::{McpConfig, McpPool, McpServerConfig, McpTimeouts};
-use serde::{Deserialize, Serialize};
-
-use crate::platform::credential_store::{
-    redact_secret, CredentialError, CredentialReference, CredentialStore, SystemCredentialStore,
-};
+use crate::platform::credential_store::{CredentialStore, SystemCredentialStore};
 use crate::platform::paths;
 
-/// 按会话类型 scope 持久化连接器禁用列表。
+// 对外 pub 面保持不变:类型从 types 子模块 re-export。
+pub use types::{
+    ConfigField, MarketplaceToolInfo, MarketplaceToolValidation, McpSecretMigrationResult,
+    RemoteOAuthConfig, RemoteServer, SecretEnv, SecretHeader, ToolManifest,
+};
+
+/// 按会话类型 scope 持久化连接器禁用列表并刷新技能目录。
 ///
 /// Refreshing live engines (组合目录重写 + 工具热刷) is an application
 /// orchestration concern and is deliberately left to the caller, keeping
@@ -36,347 +52,6 @@ pub async fn apply_disabled_connectors_for(
     .await
     .map_err(|error| format!("apply_disabled_connectors_for join: {error}"))?;
     Ok(())
-}
-
-// ---------------------------------------------------------------------------
-// Manifest — 每个 MCP 工具的元数据
-// ---------------------------------------------------------------------------
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ToolManifest {
-    pub id: String,
-    pub name: String,
-    pub description: String,
-    pub version: String,
-    pub icon: String,
-    pub category: String,
-    pub mcp_tools: Vec<String>,
-    pub command: String,
-    pub args: Vec<String>,
-    #[serde(default)]
-    pub env: std::collections::HashMap<String, String>,
-    #[serde(default)]
-    pub secret_env: Vec<SecretEnv>,
-    #[serde(default)]
-    pub secret_headers: Vec<SecretHeader>,
-    #[serde(default)]
-    pub validate_on_install: bool,
-    #[serde(default)]
-    pub config_fields: Vec<ConfigField>,
-    #[serde(default)]
-    pub routing_rules: Vec<String>,
-    #[serde(default)]
-    pub tool_table_entries: Vec<String>,
-    #[serde(default)]
-    pub pip_dependencies: Vec<String>,
-    #[serde(default)]
-    pub servers: Vec<RemoteServer>,
-    /// 配套技能 id:装该 MCP 时一并装、卸时一并删(让"一个能力"=引擎+引导整体装卸)。
-    #[serde(default)]
-    pub companion_skills: Vec<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RemoteServer {
-    pub name: String,
-    pub url: String,
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub scopes: Vec<String>,
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub oauth: Option<RemoteOAuthConfig>,
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub oauth_resource: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RemoteOAuthConfig {
-    #[serde(default)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub client_id: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SecretEnv {
-    pub key: String,
-    pub provider: String,
-    #[serde(default = "default_required")]
-    pub required: bool,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SecretHeader {
-    pub header: String,
-    #[serde(default = "default_bearer_scheme")]
-    pub scheme: String,
-    pub source_key: String,
-    pub provider: String,
-    #[serde(default = "default_required")]
-    pub required: bool,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ConfigField {
-    pub key: String,
-    pub label: String,
-    pub required: bool,
-    /// "env" = 写入 mcp.json env 字段, "bearer" = 写入 headers Authorization
-    #[serde(default = "default_target")]
-    pub target: String,
-    #[serde(default)]
-    pub secret: bool,
-}
-
-fn default_target() -> String {
-    "env".to_string()
-}
-
-fn default_required() -> bool {
-    true
-}
-
-fn default_bearer_scheme() -> String {
-    "Bearer".to_string()
-}
-
-fn is_sensitive_key_name(key: &str) -> bool {
-    let upper = key.to_ascii_uppercase();
-    upper.ends_with("_API_KEY")
-        || upper.ends_with("_TOKEN")
-        || upper.ends_with("_SECRET")
-        || upper == "API_KEY"
-        || upper == "TOKEN"
-        || upper == "SECRET"
-        || upper == "KEY"
-}
-
-fn mcp_secret_env_var(secret_name: &str) -> String {
-    let suffix = secret_name
-        .chars()
-        .map(|c| {
-            if c.is_ascii_alphanumeric() {
-                c.to_ascii_uppercase()
-            } else {
-                '_'
-            }
-        })
-        .collect::<String>();
-    format!("PINVOU3_MCP_SECRET_{suffix}")
-}
-
-fn mcp_secret_placeholder(secret_name: &str) -> String {
-    format!("${{{}}}", mcp_secret_env_var(secret_name))
-}
-
-/// 远程 MCP 的密钥不能写进 `headers`:底座会把那个字段当作字面量发送,
-/// 不会展开 `${ENV}` 占位符。Bearer 走专用的环境变量配置;无 scheme 的
-/// 自定义 header 则使用 `env_headers`。这样密钥始终只在进程环境和凭据库中。
-fn set_remote_secret_header(
-    env_headers: &mut serde_json::Map<String, serde_json::Value>,
-    bearer_token_env_var: &mut Option<String>,
-    header: &str,
-    scheme: &str,
-    key: &str,
-) -> Result<(), String> {
-    let env_var = mcp_secret_env_var(key);
-    if header.eq_ignore_ascii_case("authorization") && scheme.eq_ignore_ascii_case("bearer") {
-        if let Some(existing) = bearer_token_env_var.as_deref() {
-            if existing != env_var {
-                return Err("同一个远程 MCP server 不支持多个 Bearer 密钥".to_string());
-            }
-        }
-        *bearer_token_env_var = Some(env_var);
-        return Ok(());
-    }
-    if scheme.trim().is_empty() {
-        env_headers.insert(header.to_string(), serde_json::Value::String(env_var));
-        return Ok(());
-    }
-    Err(format!(
-        "远程 MCP 密钥 header '{header}' 的 scheme '{scheme}' 暂不支持；请使用 Bearer Authorization 或无 scheme 的自定义 header"
-    ))
-}
-
-fn write_json_pretty(path: &std::path::Path, value: &serde_json::Value) -> Result<(), String> {
-    let json = serde_json::to_string_pretty(value).map_err(|e| e.to_string())?;
-    std::fs::write(path, json).map_err(|e| format!("写入 {} 失败: {e}", path.display()))
-}
-
-fn mcp_secret_reference(tool_id: &str, target: &str, key: &str) -> CredentialReference {
-    CredentialReference::for_mcp_secret(tool_id, target, key)
-}
-
-fn mcp_secret_missing_error(tool_id: &str, key: &str) -> String {
-    format!("MCP 工具 '{tool_id}' 缺少密钥 {key}，请重新配置后再启用该工具")
-}
-
-fn mcp_secret_store_error(tool_id: &str, key: &str, error: CredentialError) -> String {
-    redact_secret(&format!(
-        "MCP 工具 '{tool_id}' 的密钥 {key} 无法访问: {}",
-        error.user_message()
-    ))
-}
-
-fn expected_remote_tool_names(manifest: &ToolManifest) -> Vec<String> {
-    if !manifest.mcp_tools.is_empty() {
-        return manifest
-            .mcp_tools
-            .iter()
-            .map(|name| normalize_manifest_tool_name(name, manifest))
-            .collect();
-    }
-    Vec::new()
-}
-
-fn normalize_manifest_tool_name(name: &str, manifest: &ToolManifest) -> String {
-    for server in &manifest.servers {
-        let prefix = format!("mcp_{}_", server.name);
-        if let Some(rest) = name.strip_prefix(&prefix) {
-            return rest.to_string();
-        }
-    }
-    let id_prefix = format!("mcp_{}_", manifest.id);
-    if let Some(rest) = name.strip_prefix(&id_prefix) {
-        return rest.to_string();
-    }
-    name.to_string()
-}
-
-fn remote_validation_user_error(raw: &str) -> String {
-    let redacted = redact_secret(raw);
-    let lower = redacted.to_ascii_lowercase();
-    let auth_failed = [
-        "401",
-        "403",
-        "unauthorized",
-        "forbidden",
-        "invalid token",
-        "invalid api key",
-        "invalid apikey",
-        "api key invalid",
-        "apikey invalid",
-        "expired",
-        "authentication failed",
-        "auth failed",
-        "permission denied",
-        "access denied",
-        "鉴权",
-        "认证",
-        "无效",
-        "过期",
-        "权限",
-    ]
-    .iter()
-    .any(|needle| lower.contains(needle));
-    let network_error = [
-        "dns",
-        "tls",
-        "certificate",
-        "connection refused",
-        "connection reset",
-        "connect error",
-        "proxy",
-        "network",
-        "failed to lookup address",
-    ]
-    .iter()
-    .any(|needle| lower.contains(needle));
-
-    if auth_failed {
-        "API Key 无效或已过期，请更新后重试".to_string()
-    } else if lower.contains("429") || lower.contains("too many requests") {
-        "远程 MCP 服务当前限流，请稍后重试".to_string()
-    } else if lower.contains("timed out") || lower.contains("timeout") || lower.contains("超时") {
-        "远程 MCP 服务响应超时，请稍后重试".to_string()
-    } else if lower.contains("tools/list") || lower.contains("工具列表") {
-        "远程 MCP 工具列表异常，请稍后重试".to_string()
-    } else if network_error {
-        "无法连接远程 MCP 服务，请检查网络或代理".to_string()
-    } else {
-        "远程 MCP 连接校验失败，请检查 API Key 或稍后重试".to_string()
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-struct LegacyMcpSecretSpec {
-    tool_id: &'static str,
-    target: &'static str,
-    key: &'static str,
-}
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
-pub struct McpSecretMigrationResult {
-    pub migrated_count: usize,
-    pub skipped_count: usize,
-    pub failed_count: usize,
-    pub messages: Vec<String>,
-}
-
-fn legacy_mcp_secret_specs() -> &'static [LegacyMcpSecretSpec] {
-    &[
-        LegacyMcpSecretSpec {
-            tool_id: "weather",
-            target: "env",
-            key: "AMAP_KEY",
-        },
-        LegacyMcpSecretSpec {
-            tool_id: "iwencai",
-            target: "env",
-            key: "IWENCAI_API_KEY",
-        },
-        LegacyMcpSecretSpec {
-            tool_id: "qcc",
-            target: "header",
-            key: "QCC_API_KEY",
-        },
-    ]
-}
-
-fn legacy_spec_for_tool(tool_id: &str) -> Option<&'static LegacyMcpSecretSpec> {
-    legacy_mcp_secret_specs()
-        .iter()
-        .find(|spec| spec.tool_id == tool_id)
-}
-
-fn legacy_spec_for_server_name(server_name: &str) -> Option<&'static LegacyMcpSecretSpec> {
-    if server_name == "weather" {
-        legacy_spec_for_tool("weather")
-    } else if server_name == "iwencai" {
-        legacy_spec_for_tool("iwencai")
-    } else if server_name.starts_with("qcc-") {
-        legacy_spec_for_tool("qcc")
-    } else {
-        None
-    }
-}
-
-// ---------------------------------------------------------------------------
-// MarketplaceToolInfo — 前端展示用
-// ---------------------------------------------------------------------------
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct MarketplaceToolInfo {
-    pub id: String,
-    pub name: String,
-    pub description: String,
-    pub version: String,
-    pub icon: String,
-    pub category: String,
-    pub installed: bool,
-    /// 配套技能 id(来自 manifest `companion_skills`)。前端据此把「有配套 MCP 的技能卡」的
-    /// 状态/装卸联动到本 MCP,单一真源在 manifest,避免命名不一致(gongwen↔government-writing)
-    /// 时前端漏建映射导致两卡状态分叉。
-    #[serde(default)]
-    pub companion_skills: Vec<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct MarketplaceToolValidation {
-    pub tool_id: String,
-    pub connected: bool,
-    pub tools: Vec<String>,
 }
 
 // ---------------------------------------------------------------------------
@@ -588,41 +263,6 @@ pub fn remove_connector_from_disabled_scopes(tool_id: &str) {
     }
 }
 
-/// 从 manifest 提取所有 secret 的 (keyring target, key):
-/// `secret_env`→("env",key)、`secret_headers`→("header",source_key)、
-/// `config_fields`(secret=true)→(env 或 header, key)。同一 (target,key) 去重一次。
-/// 与 install 时 `resolve_secret_placeholder` 用的 target 对齐(config_fields 的
-/// "bearer" 在 install 里落成 reference target "header")。
-fn manifest_secret_targets(manifest: &ToolManifest) -> Vec<(String, String)> {
-    let mut out: Vec<(String, String)> = Vec::new();
-    let mut push = |target: &str, key: &str| {
-        let pair = (target.to_string(), key.to_string());
-        if !out.contains(&pair) {
-            out.push(pair);
-        }
-    };
-    for s in &manifest.secret_env {
-        push("env", &s.key);
-    }
-    for s in &manifest.secret_headers {
-        push(
-            bundle::keyring_target(bundle::CredentialTarget::Bearer),
-            &s.source_key,
-        );
-    }
-    for f in &manifest.config_fields {
-        if f.secret {
-            let target = if f.target == "bearer" {
-                bundle::keyring_target(bundle::CredentialTarget::Bearer)
-            } else {
-                f.target.as_str()
-            };
-            push(target, &f.key);
-        }
-    }
-    out
-}
-
 /// 重启后把**所有已安装工具**的 secret 从 keyring 重灌进进程 env(MCP 子进程 expand
 /// `${...}` 占位符用)。不再硬编码内置 3 个 —— 自定义/上传的带 secret 工具重启后同样生效。
 pub fn sync_mcp_secret_env_vars() -> Result<(), String> {
@@ -640,15 +280,16 @@ pub fn disabled_tool_names_for(scope: ConnectorScope) -> Vec<String> {
 }
 
 // ---------------------------------------------------------------------------
-// MarketplaceManager
+// MarketplaceManager — facade:构造 + 核心 manifest/installed 读写 + 装卸编排。
+// secret/migration/validation/connectors 的方法分别在各子模块的 impl 块里定义。
 // ---------------------------------------------------------------------------
 
 pub struct MarketplaceManager<S: CredentialStore = SystemCredentialStore> {
     /// bundle 解包后的 MCP servers 目录 (~/.pinvou3/bundle/mcp-servers/)
-    servers_dir: PathBuf,
+    pub(crate) servers_dir: PathBuf,
     /// 已安装工具列表文件 (~/.pinvou3/marketplace/installed.json)
-    installed_file: PathBuf,
-    credential_store: S,
+    pub(crate) installed_file: PathBuf,
+    pub(crate) credential_store: S,
 }
 
 impl Default for MarketplaceManager<SystemCredentialStore> {
@@ -674,25 +315,6 @@ impl<S: CredentialStore> MarketplaceManager<S> {
             installed_file,
             credential_store,
         }
-    }
-
-    fn sync_secret_env_vars(&self) -> Result<(), String> {
-        for tool_id in self.installed_ids() {
-            let Some(manifest) = self.load_manifest(&tool_id) else {
-                continue;
-            };
-            for (target, key) in manifest_secret_targets(&manifest) {
-                let reference = mcp_secret_reference(&tool_id, &target, &key);
-                match self.credential_store.get(&reference) {
-                    Ok(Some(value)) if !value.trim().is_empty() => {
-                        std::env::set_var(mcp_secret_env_var(&key), value);
-                    }
-                    Ok(_) => {}
-                    Err(e) => return Err(mcp_secret_store_error(&tool_id, &key, e)),
-                }
-            }
-        }
-        Ok(())
     }
 
     /// 扫描 bundle mcp-servers/ 下所有含 manifest.json 的子目录
@@ -790,376 +412,15 @@ impl<S: CredentialStore> MarketplaceManager<S> {
         Ok(())
     }
 
-    /// manifest 显式要求时，把“配置已写入”收紧为“远程 MCP 已握手且工具可发现”。
-    pub fn requires_remote_connection_validation(&self, tool_id: &str) -> bool {
-        self.load_manifest(tool_id)
-            .map(|m| m.validate_on_install && !m.servers.is_empty())
-            .unwrap_or(false)
-    }
-
-    pub async fn validate_remote_connection(
-        &self,
-        tool_id: &str,
-    ) -> Result<MarketplaceToolValidation, String> {
-        let manifest = self
-            .load_manifest(tool_id)
-            .ok_or_else(|| format!("工具 '{tool_id}' 不存在"))?;
-        if !self.requires_remote_connection_validation(tool_id) {
-            return Ok(MarketplaceToolValidation {
-                tool_id: tool_id.to_string(),
-                connected: true,
-                tools: Vec::new(),
-            });
-        }
-
-        let mut pool = McpPool::new(self.validation_mcp_config(&manifest)?);
-        let errors = pool.connect_all().await;
-        if !errors.is_empty() {
-            let message = errors
-                .into_iter()
-                .map(|(server, err)| format!("{server}: {err:#}"))
-                .collect::<Vec<_>>()
-                .join("; ");
-            return Err(remote_validation_user_error(&message));
-        }
-
-        let tools = pool
-            .all_tools()
-            .into_iter()
-            .map(|(_, tool)| tool.name.clone())
-            .collect::<Vec<_>>();
-        if tools.is_empty() {
-            return Err("远程 MCP 工具列表异常，请稍后重试".to_string());
-        }
-
-        let expected = expected_remote_tool_names(&manifest);
-        if !expected.is_empty() {
-            let missing = expected
-                .iter()
-                .filter(|name| !tools.iter().any(|tool| tool == *name))
-                .cloned()
-                .collect::<Vec<_>>();
-            if !missing.is_empty() {
-                return Err(format!(
-                    "远程 MCP 工具列表异常，缺少工具: {}",
-                    missing.join(", ")
-                ));
-            }
-        }
-
-        Ok(MarketplaceToolValidation {
-            tool_id: tool_id.to_string(),
-            connected: true,
-            tools,
-        })
-    }
-
-    fn validation_mcp_config(&self, manifest: &ToolManifest) -> Result<McpConfig, String> {
-        let mcp_path = paths::mcp_config_path();
-        let content =
-            std::fs::read_to_string(&mcp_path).map_err(|e| format!("读取 MCP 配置失败: {e}"))?;
-        let mcp: serde_json::Value =
-            serde_json::from_str(&content).map_err(|e| format!("解析 MCP 配置失败: {e}"))?;
-        let servers = mcp
-            .get("servers")
-            .and_then(|v| v.as_object())
-            .ok_or_else(|| "MCP 配置缺少 servers".to_string())?;
-
-        let mut config = McpConfig {
-            timeouts: McpTimeouts {
-                connect_timeout: 20,
-                execute_timeout: 30,
-                read_timeout: 30,
-            },
-            servers: std::collections::HashMap::new(),
-        };
-        for server in &manifest.servers {
-            let entry = servers
-                .get(&server.name)
-                .ok_or_else(|| format!("MCP 配置缺少 server '{}'", server.name))?;
-            let mut server_config: McpServerConfig = serde_json::from_value(entry.clone())
-                .map_err(|e| format!("解析 MCP server '{}' 失败: {e}", server.name))?;
-            server_config.required = true;
-            server_config.connect_timeout = Some(20);
-            server_config.execute_timeout = Some(30);
-            server_config.read_timeout = Some(30);
-            config.servers.insert(server.name.clone(), server_config);
-        }
-        Ok(config)
-    }
-
-    pub fn migrate_mcp_plaintext_secrets(&self) -> Result<McpSecretMigrationResult, String> {
-        let mut result = McpSecretMigrationResult::default();
-        for spec in legacy_mcp_secret_specs() {
-            let path = self.servers_dir.join(spec.tool_id).join("manifest.json");
-            if path.is_file() {
-                self.migrate_manifest_file(&path, spec, &mut result)?;
-            }
-        }
-        let mcp_path = paths::mcp_config_path();
-        if mcp_path.is_file() {
-            self.migrate_mcp_json_file(&mcp_path, &mut result)?;
-        }
-        Ok(result)
-    }
-
-    /// 装 `manifest.pip_dependencies` 里的 Python 依赖（跨平台）。
-    /// 用 `python -m pip install`（保证装进跑 MCP server 的同一个 python，不裸 `pip`）。
-    /// ① 先预检依赖是否已可用（系统已装/此前装过）→ 命中即跳过，不跑 pip；
-    /// ② 否则按序兜底：`--user` → `--user --break-system-packages`（PEP 668）→ `--break-system-packages`，任一成功即 Ok。
-    /// 零依赖工具（pip_dependencies 为空）直接返回 Ok，不影响 weather/obsidian 等。
-    fn pip_install_deps(&self, manifest: &ToolManifest) -> Result<(), String> {
-        if manifest.pip_dependencies.is_empty() {
-            return Ok(());
-        }
-        // Windows:python-pptx 等依赖已随内置 python(python-win)预装,不在用户机器
-        // 跑 pip —— 用户也就不需要自己装 python。仅 Linux/macOS 走系统 python3 联网 pip。
-        if crate::platform::capabilities::is_windows() {
-            return Ok(());
-        }
-        let python_cmd = "python3";
-        let deps = &manifest.pip_dependencies;
-
-        // ① 预检:依赖已可用就直接 Ok,不跑 pip。用 importlib.metadata 按 PyPI 包名查
-        //    (python-pptx 等),全部命中即满足。修「明明已装(系统包/此前装过)却仍判失败」。
-        let satisfied = std::process::Command::new(python_cmd)
-            .arg("-c")
-            .arg("import importlib.metadata as m, sys; [m.version(p) for p in sys.argv[1:]]")
-            .args(deps)
-            .output()
-            .map(|o| o.status.success())
-            .unwrap_or(false);
-        if satisfied {
-            return Ok(());
-        }
-
-        // ② pip 安装,按序兜底,任一成功即 Ok:
-        //    --user(常规)→ --user --break-system-packages(PEP 668:现代 Debian/Ubuntu 拦 --user,
-        //    装进 ~/.local 用户目录、不动系统/发行版包)→ --break-system-packages(某些环境 --user 不可用)。
-        let run = |extra: &[&str]| -> std::io::Result<std::process::Output> {
-            let mut cmd = std::process::Command::new(python_cmd);
-            cmd.args([
-                "-m",
-                "pip",
-                "install",
-                "--disable-pip-version-check",
-                "--no-input",
-            ]);
-            cmd.args(extra);
-            cmd.args(deps);
-            cmd.output()
-        };
-        let attempts: [&[&str]; 3] = [
-            &["--user"],
-            &["--user", "--break-system-packages"],
-            &["--break-system-packages"],
-        ];
-        let mut last_err = String::new();
-        for extra in attempts {
-            match run(extra) {
-                Ok(o) if o.status.success() => return Ok(()),
-                Ok(o) => {
-                    last_err = String::from_utf8_lossy(&o.stderr)
-                        .trim()
-                        .lines()
-                        .last()
-                        .unwrap_or("")
-                        .to_string();
-                }
-                Err(e) => {
-                    return Err(format!(
-                        "无法运行 {python_cmd}（请确认已安装 Python 且在 PATH 中）：{e}"
-                    ));
-                }
-            }
-        }
-        Err(format!(
-            "依赖安装失败（pip）：{last_err}（已尝试 --user 与 --break-system-packages;请确认网络可达且 python3 自带 pip）"
-        ))
-    }
-
-    fn resolve_secret_placeholder(
-        &self,
-        tool_id: &str,
-        target: &str,
-        key: &str,
-        user_config: &std::collections::HashMap<String, String>,
-        legacy_env: &std::collections::HashMap<String, String>,
-    ) -> Result<String, String> {
-        let reference = mcp_secret_reference(tool_id, target, key);
-        if let Some(value) = user_config.get(key).filter(|v| !v.trim().is_empty()) {
-            self.credential_store
-                .set(&reference, value)
-                .map_err(|e| mcp_secret_store_error(tool_id, key, e))?;
-            std::env::set_var(mcp_secret_env_var(key), value);
-            return Ok(mcp_secret_placeholder(key));
-        }
-
-        match self.credential_store.get(&reference) {
-            Ok(Some(value)) if !value.trim().is_empty() => {
-                std::env::set_var(mcp_secret_env_var(key), value);
-                Ok(mcp_secret_placeholder(key))
-            }
-            Ok(_) => {
-                if let Some(value) = legacy_env.get(key).filter(|v| !v.trim().is_empty()) {
-                    self.credential_store
-                        .set(&reference, value)
-                        .map_err(|e| mcp_secret_store_error(tool_id, key, e))?;
-                    std::env::set_var(mcp_secret_env_var(key), value);
-                    Ok(mcp_secret_placeholder(key))
-                } else {
-                    Err(mcp_secret_missing_error(tool_id, key))
-                }
-            }
-            Err(e) => Err(mcp_secret_store_error(tool_id, key, e)),
-        }
-    }
-
-    fn migrate_manifest_file(
-        &self,
-        path: &std::path::Path,
-        spec: &LegacyMcpSecretSpec,
-        result: &mut McpSecretMigrationResult,
-    ) -> Result<(), String> {
-        let content = std::fs::read_to_string(path)
-            .map_err(|e| format!("读取 {} 失败: {e}", path.display()))?;
-        let mut json: serde_json::Value = serde_json::from_str(&content)
-            .map_err(|e| format!("解析 {} 失败: {e}", path.display()))?;
-        let Some(value) = json
-            .get("env")
-            .and_then(|env| env.get(spec.key))
-            .and_then(|v| v.as_str())
-            .filter(|v| !v.trim().is_empty())
-            .map(ToOwned::to_owned)
-        else {
-            return Ok(());
-        };
-
-        self.store_migrated_secret(spec, &value, result)?;
-        if let Some(env) = json.get_mut("env").and_then(|env| env.as_object_mut()) {
-            env.remove(spec.key);
-            if env.is_empty() {
-                json.as_object_mut().map(|obj| obj.remove("env"));
-            }
-        }
-        write_json_pretty(path, &json)?;
-        Ok(())
-    }
-
-    fn migrate_mcp_json_file(
-        &self,
-        path: &std::path::Path,
-        result: &mut McpSecretMigrationResult,
-    ) -> Result<(), String> {
-        let content = std::fs::read_to_string(path)
-            .map_err(|e| format!("读取 {} 失败: {e}", path.display()))?;
-        let mut json: serde_json::Value = serde_json::from_str(&content)
-            .map_err(|e| format!("解析 {} 失败: {e}", path.display()))?;
-        let mut changed = false;
-        let Some(servers) = json
-            .get_mut("servers")
-            .and_then(|servers| servers.as_object_mut())
-        else {
-            return Ok(());
-        };
-
-        for (server_name, entry) in servers.iter_mut() {
-            let Some(spec) = legacy_spec_for_server_name(server_name) else {
-                continue;
-            };
-            if let Some(env) = entry.get_mut("env").and_then(|env| env.as_object_mut()) {
-                if let Some(value) = env
-                    .get(spec.key)
-                    .and_then(|v| v.as_str())
-                    .filter(|v| !v.trim().is_empty())
-                    .map(ToOwned::to_owned)
-                {
-                    self.store_migrated_secret(spec, &value, result)?;
-                    env.insert(
-                        spec.key.to_string(),
-                        serde_json::Value::String(mcp_secret_placeholder(spec.key)),
-                    );
-                    changed = true;
-                }
-            }
-            if let Some(headers) = entry
-                .get_mut("headers")
-                .and_then(|headers| headers.as_object_mut())
-            {
-                if let Some(auth) = headers
-                    .get("Authorization")
-                    .and_then(|v| v.as_str())
-                    .filter(|v| !v.trim().is_empty())
-                    .map(ToOwned::to_owned)
-                {
-                    if let Some(secret) = auth.strip_prefix("Bearer ").filter(|v| !v.is_empty()) {
-                        self.store_migrated_secret(spec, secret, result)?;
-                        // `headers` 是字面量,不会展开 `${ENV}`。迁移到
-                        // 底座的 Bearer 环境变量字段,避免“已迁移但实际鉴权失败”。
-                        headers.remove("Authorization");
-                        if headers.is_empty() {
-                            entry.as_object_mut().map(|object| object.remove("headers"));
-                        }
-                        entry["bearer_token_env_var"] =
-                            serde_json::Value::String(mcp_secret_env_var(spec.key));
-                        changed = true;
-                    }
-                }
-            }
-        }
-
-        if changed {
-            write_json_pretty(path, &json)?;
-        }
-        Ok(())
-    }
-
-    fn store_migrated_secret(
-        &self,
-        spec: &LegacyMcpSecretSpec,
-        value: &str,
-        result: &mut McpSecretMigrationResult,
-    ) -> Result<(), String> {
-        let reference = mcp_secret_reference(spec.tool_id, spec.target, spec.key);
-        let env_value = match self.credential_store.get(&reference) {
-            Ok(Some(existing)) if !existing.trim().is_empty() => {
-                result.skipped_count += 1;
-                result.messages.push(format!(
-                    "MCP 工具 '{}' 的密钥 {} 已存在，已跳过覆盖并清理旧明文",
-                    spec.tool_id, spec.key
-                ));
-                existing
-            }
-            Ok(_) => {
-                self.credential_store.set(&reference, value).map_err(|e| {
-                    result.failed_count += 1;
-                    mcp_secret_store_error(spec.tool_id, spec.key, e)
-                })?;
-                result.migrated_count += 1;
-                result.messages.push(format!(
-                    "MCP 工具 '{}' 的密钥 {} 已迁移到系统凭据存储",
-                    spec.tool_id, spec.key
-                ));
-                value.to_string()
-            }
-            Err(e) => {
-                result.failed_count += 1;
-                return Err(mcp_secret_store_error(spec.tool_id, spec.key, e));
-            }
-        };
-        std::env::set_var(mcp_secret_env_var(spec.key), env_value);
-        Ok(())
-    }
-
     /// 卸载工具：从 installed.json + mcp.json 中移除
     pub fn uninstall(&self, tool_id: &str) -> Result<(), String> {
         // 删该工具在 keyring 的 secret(防孤儿;此时 manifest 未删、仍可读声明)。
         // 删不掉不阻断卸载；若用户重新安装并重新填 key，会写入新的系统凭据。
         if let Some(manifest) = self.load_manifest(tool_id) {
-            for (target, key) in manifest_secret_targets(&manifest) {
-                let reference = mcp_secret_reference(tool_id, &target, &key);
+            for (target, key) in secrets::manifest_secret_targets(&manifest) {
+                let reference = secrets::mcp_secret_reference(tool_id, &target, &key);
                 let _ = self.credential_store.delete(&reference);
-                std::env::remove_var(mcp_secret_env_var(&key));
+                std::env::remove_var(secrets::mcp_secret_env_var(&key));
             }
         }
         // 更新 installed.json
@@ -1325,237 +586,6 @@ impl<S: CredentialStore> MarketplaceManager<S> {
         }
         recovered
     }
-
-    fn add_to_mcp_json(
-        &self,
-        manifest: &ToolManifest,
-        user_config: &std::collections::HashMap<String, String>,
-    ) -> Result<(), String> {
-        let mcp_path = paths::mcp_config_path();
-        let mut mcp: serde_json::Value = if mcp_path.is_file() {
-            let content =
-                std::fs::read_to_string(&mcp_path).map_err(|e| format!("读取 mcp.json: {e}"))?;
-            serde_json::from_str(&content).unwrap_or_else(|_| default_mcp_json())
-        } else {
-            default_mcp_json()
-        };
-
-        let servers = mcp
-            .get_mut("servers")
-            .and_then(|s| s.as_object_mut())
-            .ok_or("mcp.json 格式错误")?;
-
-        if !manifest.servers.is_empty() {
-            // ── 远程工具：遍历 manifest.servers[]，写 url/headers ──
-            for server in &manifest.servers {
-                let mut headers = serde_json::Map::new();
-                let mut env_headers = serde_json::Map::new();
-                let mut bearer_token_env_var = None;
-
-                // 1. config_fields 中 target="bearer" 的字段（用户填入）
-                for field in &manifest.config_fields {
-                    if field.target == "bearer" {
-                        if let Some(val) = user_config.get(&field.key) {
-                            if field.secret {
-                                self.resolve_secret_placeholder(
-                                    &manifest.id,
-                                    bundle::keyring_target(bundle::CredentialTarget::Bearer),
-                                    &field.key,
-                                    user_config,
-                                    &manifest.env,
-                                )?;
-                                set_remote_secret_header(
-                                    &mut env_headers,
-                                    &mut bearer_token_env_var,
-                                    "Authorization",
-                                    "Bearer",
-                                    &field.key,
-                                )?;
-                            } else {
-                                headers.insert(
-                                    "Authorization".to_string(),
-                                    serde_json::Value::String(format!("Bearer {}", val)),
-                                );
-                            }
-                        }
-                    }
-                }
-
-                // 2. manifest.secret_headers 声明的敏感 header（不落明文）
-                for secret in &manifest.secret_headers {
-                    self.resolve_secret_placeholder(
-                        &manifest.id,
-                        bundle::keyring_target(bundle::CredentialTarget::Bearer),
-                        &secret.source_key,
-                        user_config,
-                        &manifest.env,
-                    )?;
-                    set_remote_secret_header(
-                        &mut env_headers,
-                        &mut bearer_token_env_var,
-                        &secret.header,
-                        &secret.scheme,
-                        &secret.source_key,
-                    )?;
-                }
-
-                // 3. 兼容旧 manifest.env 中以 _API_KEY 结尾的字段，迁移后只写占位。
-                if headers.is_empty() {
-                    for k in manifest.env.keys() {
-                        if is_sensitive_key_name(k) {
-                            let placeholder = self.resolve_secret_placeholder(
-                                &manifest.id,
-                                bundle::keyring_target(bundle::CredentialTarget::Bearer),
-                                k,
-                                user_config,
-                                &manifest.env,
-                            )?;
-                            headers.insert(
-                                "Authorization".to_string(),
-                                serde_json::Value::String(format!("Bearer {}", placeholder)),
-                            );
-                            break;
-                        }
-                    }
-                }
-
-                let mut entry = serde_json::json!({ "url": server.url });
-                if !server.scopes.is_empty() {
-                    entry["scopes"] = serde_json::to_value(&server.scopes).unwrap_or_default();
-                }
-                if let Some(oauth) = &server.oauth {
-                    entry["oauth"] = serde_json::to_value(oauth).unwrap_or_default();
-                }
-                if let Some(resource) = &server.oauth_resource {
-                    if !resource.trim().is_empty() {
-                        entry["oauth_resource"] = serde_json::Value::String(resource.clone());
-                    }
-                }
-                if !headers.is_empty() {
-                    entry["headers"] = serde_json::Value::Object(headers);
-                }
-                if !env_headers.is_empty() {
-                    entry["env_headers"] = serde_json::Value::Object(env_headers);
-                }
-                if let Some(env_var) = bearer_token_env_var {
-                    entry["bearer_token_env_var"] = serde_json::Value::String(env_var);
-                }
-                servers.insert(server.name.clone(), entry);
-            }
-        } else {
-            // ── 本地工具：command/args/env ──
-            let server_dir = self.servers_dir.join(&manifest.id);
-            let args: Vec<String> = manifest
-                .args
-                .iter()
-                .map(|a| {
-                    if a == "server.py" || a.ends_with("/server.py") {
-                        server_dir.join("server.py").to_string_lossy().to_string()
-                    } else {
-                        a.clone()
-                    }
-                })
-                .collect();
-
-            let mut env = manifest
-                .env
-                .iter()
-                .filter(|(k, _)| !is_sensitive_key_name(k))
-                .map(|(k, v)| (k.clone(), v.clone()))
-                .collect::<std::collections::HashMap<_, _>>();
-            for field in &manifest.config_fields {
-                if field.target == "env" {
-                    if let Some(val) = user_config.get(&field.key) {
-                        if field.secret || is_sensitive_key_name(&field.key) {
-                            let placeholder = self.resolve_secret_placeholder(
-                                &manifest.id,
-                                "env",
-                                &field.key,
-                                user_config,
-                                &manifest.env,
-                            )?;
-                            env.insert(field.key.clone(), placeholder);
-                        } else {
-                            env.insert(field.key.clone(), val.clone());
-                        }
-                    }
-                }
-            }
-            for secret in &manifest.secret_env {
-                let placeholder = self.resolve_secret_placeholder(
-                    &manifest.id,
-                    "env",
-                    &secret.key,
-                    user_config,
-                    &manifest.env,
-                )?;
-                env.insert(secret.key.clone(), placeholder);
-            }
-            for key in manifest.env.keys().filter(|k| is_sensitive_key_name(k)) {
-                if !env.contains_key(key) {
-                    let placeholder = self.resolve_secret_placeholder(
-                        &manifest.id,
-                        "env",
-                        key,
-                        user_config,
-                        &manifest.env,
-                    )?;
-                    env.insert(key.clone(), placeholder);
-                }
-            }
-
-            // python 工具:Windows 用内置 pythonw(无窗口 + 自带依赖),其他平台系统 python3。
-            let command = if manifest.command == "python" || manifest.command == "python3" {
-                paths::python_command()
-            } else {
-                manifest.command.clone()
-            };
-            let mut entry = serde_json::json!({
-                "command": command,
-                "args": args,
-            });
-            if !env.is_empty() {
-                entry["env"] = serde_json::to_value(&env).unwrap_or_default();
-            }
-
-            servers.insert(manifest.id.clone(), entry);
-        }
-
-        write_json_pretty(&mcp_path, &mcp)
-    }
-
-    fn remove_from_mcp_json(&self, tool_id: &str) -> Result<(), String> {
-        let mcp_path = paths::mcp_config_path();
-        if !mcp_path.is_file() {
-            return Ok(());
-        }
-        let content =
-            std::fs::read_to_string(&mcp_path).map_err(|e| format!("读取 mcp.json: {e}"))?;
-        let mut mcp: serde_json::Value =
-            serde_json::from_str(&content).unwrap_or_else(|_| default_mcp_json());
-
-        if let Some(servers) = mcp.get_mut("servers").and_then(|s| s.as_object_mut()) {
-            // 先尝试加载 manifest 看是否有 servers 字段（远程工具有多条目）
-            if let Some(manifest) = self.load_manifest(tool_id) {
-                if !manifest.servers.is_empty() {
-                    for server in &manifest.servers {
-                        servers.remove(&server.name);
-                    }
-                } else {
-                    servers.remove(tool_id);
-                }
-            } else {
-                servers.remove(tool_id);
-            }
-        }
-
-        let json = serde_json::to_string_pretty(&mcp).map_err(|e| e.to_string())?;
-        std::fs::write(&mcp_path, json).map_err(|e| format!("写入 mcp.json: {e}"))
-    }
-}
-
-fn default_mcp_json() -> serde_json::Value {
-    serde_json::json!({"servers": {}})
 }
 
 #[cfg(test)]
@@ -1565,117 +595,12 @@ mod tests {
     use super::*;
     use crate::platform::credential_store::{CredentialStore, MemoryCredentialStore};
     use crate::platform::paths::tests::ENV_LOCK;
+    use secrets::mcp_secret_reference;
     use std::future::Future;
     use std::sync::{Arc, Mutex as StdMutex};
     use tokio::io::{AsyncReadExt, AsyncWriteExt};
     use tokio::net::TcpListener;
     use tokio::task::JoinHandle;
-
-    #[test]
-    fn manifest_secret_targets_dedups_and_maps_bearer_to_header() {
-        // 同一 key 在 secret_env/secret_headers 与 config_fields 重复声明 → 去重一次。
-        let manifest: ToolManifest = serde_json::from_str(
-            r#"{
-            "id":"t","name":"T","description":"","version":"1","icon":"","category":"",
-            "mcp_tools":[],"command":"","args":[],
-            "secret_env":[{"key":"AMAP_KEY","provider":"amap","required":true}],
-            "secret_headers":[{"header":"Authorization","scheme":"Bearer","source_key":"QCC_API_KEY","provider":"qcc","required":true}],
-            "config_fields":[
-                {"key":"AMAP_KEY","label":"","required":false,"target":"env","secret":true},
-                {"key":"QCC_API_KEY","label":"","required":false,"target":"bearer","secret":true}
-            ]
-        }"#,
-        )
-        .unwrap();
-        let targets = manifest_secret_targets(&manifest);
-        assert_eq!(targets.len(), 2, "AMAP/QCC 各去重一次");
-        assert!(targets.contains(&("env".to_string(), "AMAP_KEY".to_string())));
-        assert!(targets.contains(&("header".to_string(), "QCC_API_KEY".to_string())));
-    }
-
-    #[test]
-    fn remote_validation_error_classifier_prefers_auth_message_for_token_failures() {
-        for raw in [
-            "401 Unauthorized",
-            "403 Forbidden",
-            "invalid apikey",
-            "invalid api key",
-            "invalid token",
-            "api key invalid",
-            "authentication failed",
-            "auth failed",
-            "permission denied",
-            "access denied",
-            "鉴权失败",
-            "认证失败",
-            "API Key 无效",
-            "token expired",
-        ] {
-            assert_eq!(
-                remote_validation_user_error(raw),
-                "API Key 无效或已过期，请更新后重试",
-                "raw={raw}"
-            );
-        }
-    }
-
-    #[test]
-    fn remote_validation_error_classifier_separates_network_and_unknown_errors() {
-        for raw in [
-            "connection refused",
-            "dns lookup failed",
-            "proxy connect failed",
-            "TLS certificate error",
-            "failed to lookup address information",
-        ] {
-            assert_eq!(
-                remote_validation_user_error(raw),
-                "无法连接远程 MCP 服务，请检查网络或代理",
-                "raw={raw}"
-            );
-        }
-
-        assert_eq!(
-            remote_validation_user_error("upstream rejected request"),
-            "远程 MCP 连接校验失败，请检查 API Key 或稍后重试"
-        );
-        assert_eq!(
-            remote_validation_user_error("unexpected json-rpc error wrong-token-20260715"),
-            "远程 MCP 连接校验失败，请检查 API Key 或稍后重试"
-        );
-    }
-
-    #[test]
-    fn remote_secret_header_config_uses_environment_backed_fields() {
-        let mut env_headers = serde_json::Map::new();
-        let mut bearer_token_env_var = None;
-
-        set_remote_secret_header(
-            &mut env_headers,
-            &mut bearer_token_env_var,
-            "Authorization",
-            "Bearer",
-            "PATSNAP_API_KEY",
-        )
-        .unwrap();
-        set_remote_secret_header(
-            &mut env_headers,
-            &mut bearer_token_env_var,
-            "X-Api-Key",
-            "",
-            "EXAMPLE_API_KEY",
-        )
-        .unwrap();
-
-        assert_eq!(
-            bearer_token_env_var.as_deref(),
-            Some("PINVOU3_MCP_SECRET_PATSNAP_API_KEY")
-        );
-        assert_eq!(
-            env_headers["X-Api-Key"],
-            "PINVOU3_MCP_SECRET_EXAMPLE_API_KEY"
-        );
-    }
 
     /// 把 PINVOU3_HOME 指到一个干净临时目录跑闭包,跑完恢复并清理。
     /// 借 paths 的 ENV_LOCK 跟其它 mutate PINVOU3_HOME 的测试串行,避免互相覆盖。
@@ -1904,15 +829,6 @@ mod tests {
         let dir = crate::platform::paths::bundle_mcp_servers_dir().join(tool_id);
         std::fs::create_dir_all(&dir).unwrap();
         std::fs::write(dir.join("manifest.json"), manifest).unwrap();
-    }
-
-    /// 直接写 installed.json 模拟已安装连接器(避免走完整 install 的远程校验)。
-    fn write_installed_ids(ids: &[String]) {
-        let path = crate::platform::paths::pinvou3_home()
-            .join("marketplace")
-            .join("installed.json");
-        std::fs::create_dir_all(path.parent().unwrap()).unwrap();
-        std::fs::write(&path, serde_json::to_string(ids).unwrap()).unwrap();
     }
 
     fn read_mcp_json() -> serde_json::Value {
@@ -2213,7 +1129,14 @@ mod tests {
         });
     }
 
-    /// 双 scope(plain/code)独立持久化:互不影响,code 首次写会标记已初始化。
+    fn write_installed_ids(ids: &[String]) {
+        let path = crate::platform::paths::pinvou3_home()
+            .join("marketplace")
+            .join("installed.json");
+        std::fs::create_dir_all(path.parent().unwrap()).unwrap();
+        std::fs::write(&path, serde_json::to_string(ids).unwrap()).unwrap();
+    }
+
     #[test]
     fn disabled_connectors_scope_isolation() {
         with_temp_home(|| {
@@ -2247,7 +1170,6 @@ mod tests {
         });
     }
 
-    /// 旧版裸数组格式 `["a","b"]` 迁移到 plain scope,code 保持未初始化默认。
     #[test]
     fn disabled_connectors_legacy_array_migrates_to_plain() {
         with_temp_home(|| {
@@ -2395,7 +1317,6 @@ mod tests {
         });
     }
 
-    /// 新装连接器:code 未初始化时无需落盘(load 已默认全禁);已初始化时自动加入禁用集(默认仍关)。
     #[test]
     fn sync_deny_all_scopes_after_install_keeps_new_connector_disabled_by_default() {
         with_temp_home(|| {
@@ -2423,7 +1344,6 @@ mod tests {
         });
     }
 
-    /// 卸载连接器:从 plain/code 两个 scope 禁用集移除,避免残留 id。
     #[test]
     fn remove_connector_cleans_both_scopes() {
         with_temp_home(|| {
